Set up API
Created serializers to serialize the data for the two models. Then created json views for both models, as well as api urls to allow them to be called. Added to each model to be displayed by route_num and name in the admin. Also added to project urls to include the api urls. Added settings for the rest framework api.