Added Templates and Content
Added templates for all 8 individual routes as well as a template to show all the routes. Added content to base (to be inherited from), as well as content to main (styling, nav bar, etc). Connected all templates with individual views and urls. Added content to the all routes template to serve as a place holder and build off of for the other route templates.